requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-dotenv==1.0.0
openai==1.10.0
httpx==0.26.0
beautifulsoup4==4.12.3
aiosqlite==0.19.0
python-multipart==0.0.6
